fix: filter out invalid pools, correct poolId validity on add/show LP
fix: filter out invalid pools
