+++ src/components/pages/add-liquidity-page/components/AddLiquidity/AddLiquidity.tsx
@@ -1,5 +1,4 @@
 import styles from './AddLiquidity.module.css';
-import {createPoolIdFromPoolKey} from "@/src/utils/common";
 import {useCallback, useState} from "react";
 import PreviewAddLiquidityDialog, {AddLiquidityPreviewData} from "@/src/components/pages/add-liquidity-page/components/AddLiquidity/PreviewAddLiquidityDialog";
 import AddLiquidityDialog from "@/src/components/pages/add-liquidity-page/components/AddLiquidity/AddLiquidityDialog";
@@ -7,17 +6,16 @@ import BackLink from "@/src/components/common/BackLink/BackLink";
 import {useRouter} from "next/navigation";
 import IconButton from "@/src/components/common/IconButton/IconButton";
 import CloseIcon from "@/src/components/icons/Close/CloseIcon";
+import {PoolId} from "mira-dex-ts";
 
 type Props = {
-  poolKey: string;
+  poolId: PoolId;
 }
 
-const AddLiquidity = ({ poolKey }: Props) => {
+const AddLiquidity = ({ poolId }: Props) => {
   const router = useRouter();
 
   const [previewData, setPreviewData] = useState<AddLiquidityPreviewData | null>(null);
-
-  const poolId = createPoolIdFromPoolKey(poolKey);
 
   const handleBackClick = useCallback(() => {
     if (previewData) {
