+++ src/components/pages/view-position-page/components/PositionView/PositionView.module.css
@@ -1,0 +1,224 @@
+.viewPositionLayout {
+  display: flex;
+  flex-direction: column;
+  padding: 16px;
+  gap: 16px;
+}
+
+.contentSection {
+  display: flex;
+  flex-direction: column;
+  gap: 12px;
+}
+
+.contentSection > div:not:last-child {
+  flex: 1;
+}
+
+.coinPairAndLabel {
+  display: flex;
+  align-items: flex-start;
+  justify-content: space-between;
+}
+
+.infoBlock {
+  display: flex;
+  flex-direction: column;
+  gap: 16px;
+  padding: 12px;
+  border-radius: 16px;
+  background-color: var(--background-grey-dark);
+}
+
+.pending {
+  color: var(--content-dimmed-light);
+}
+
+.feesTitle {
+  display: flex;
+  justify-content: space-between;
+}
+
+.collectButton {
+  background: transparent;
+  color: var(--accent-primary);
+  border: none;
+  cursor: pointer;
+}
+
+.coinsData {
+  display: flex;
+  justify-content: space-between;
+}
+
+.miraBlock {
+  height: 160px;
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-end;
+  padding: 16px;
+  border-radius: 16px;
+  background: linear-gradient(359.71deg, #8350FD -1.52%, #B2EA12 121.21%);
+}
+
+.miraLogo {
+  width: 64px;
+  height: 32px;
+  margin-bottom: 12px;
+}
+
+.tokenDisplayValue, .numberAndCopy {
+  font-size: 14px;
+}
+
+.numberAndCopy {
+  display: flex;
+  justify-content: space-between;
+}
+
+.numberAndCopy svg {
+  width: 16px;
+  height: 16px;
+}
+
+.copyButton {
+  display: flex;
+  gap: 4px;
+  align-items: center;
+  color: var(--content-primary);
+  background: transparent;
+  border: none;
+  cursor: pointer;
+}
+
+.priceBlocks {
+  display: flex;
+  flex-direction: column;
+  gap: 12px;
+}
+
+.priceBlock {
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+  padding: 12px;
+  border-radius: 8px;
+  background-color: var(--background-grey-dark);
+}
+
+.priceBlockTop {
+  width: 100%;
+}
+
+.priceBlockTitle, .priceBlockDescription {
+  font-size: 14px;
+  line-height: 16px;
+  color: var(--content-dimmed-light);
+}
+
+.priceBlockValue {
+  font-size: 20px;
+  line-height: 24px;
+  color: var(--content-primary);
+}
+
+.priceBlockValueSmall {
+  font-size: 16px;
+  line-height: 19px;
+}
+
+.bottomPriceBlocks {
+  display: flex;
+  gap: 8px;
+}
+
+.bottomPriceBlocks > div {
+  width: 50%;
+}
+
+.sticky {
+  position: sticky;
+  bottom: 0;
+  width: 100%;
+  align-self: flex-start;
+}
+
+.withdrawLiquidityTitle {
+  text-align: center;
+}
+
+@media (min-width: 1024px) {
+  .viewPositionLayout {
+    /* min-height: calc(100vh - 157px); */
+    min-height: 100vh;
+    max-width: 716px;
+    margin: 0 auto;
+    padding: 32px 16px;
+
+  }
+
+  .positionHeading {
+    justify-content: space-between;
+    align-items: center;
+    display: flex;
+  }
+
+  .coinPairAndLabel {
+    gap: 8px;
+  }
+
+  p.smallLabel {
+    padding: 4px 8px;
+    font-size: 13px;
+    line-height: 15px;
+  }
+
+  .withdrawButton {
+    max-width: 191px;
+    padding: 8px 24px;
+    font-size: 13px;
+    line-height: 15px;
+  }
+
+  .topRow {
+    display: flex;
+    gap: 12px;
+    width: 100%;
+  }
+
+  .miraBlock {
+    height: unset;
+    flex: 1;
+    width: 346px;
+  }
+
+  .infoBlocks {
+    flex: 1;
+    width: 100%;
+    flex-direction: column;
+  }
+
+  .priceBlocksDesktop {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+  }
+
+  .priceBlockLargeDesktop {
+    width: 100%;
+    padding: 16px;
+    border-radius: 12px;
+    display: flex;
+    flex-direction: column;
+    gap: 16px;
+    background-color: var(--background-grey-dark);
+  }
+
+  .priceBlockDesktop {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    gap: 8px;
+  }
+}
