+++ src/utils/common.ts
@@ -1,19 +1,17 @@
 import {CoinName, coinsConfig} from "@/src/utils/coinsConfig";
 import {B256Address} from "fuels";
 import {buildPoolId, PoolId} from "mira-dex-ts";
-import {AssetIdInput} from "mira-dex-ts/dist/sdk/typegen/MiraAmmContract";
 
 export const openNewTab = (url: string) => {
   window.open(url, '_blank');
 };
 
-export const getCoinsFromKey = (key: string) => {
-  const [coinA, coinB] = key.split('-') as [CoinName, CoinName];
-  return { coinA , coinB };
-};
-
 export const getAssetNameByAssetId = (assetId: B256Address) => {
   return Array.from(coinsConfig.values()).find(coin => coin.assetId === assetId)?.name!;
+};
+
+export const isPoolIdValid = (poolId: PoolId) => {
+  return coinsConfig.has(getAssetNameByAssetId(poolId[0].bits)) && coinsConfig.has(getAssetNameByAssetId(poolId[1].bits));
 };
 
 export const getAssetNamesFromPoolId = (poolId: PoolId) => {
@@ -33,6 +31,11 @@ export const createPoolKey = (poolId: PoolId) => {
   return `${firstAssetName}-${secondAssetName}-${poolStability}`;
 };
 
+export const getCoinsFromKey = (key: string) => {
+  const [coinA, coinB] = key.split('-') as [CoinName, CoinName];
+  return { coinA , coinB };
+};
+
 // TODO: Reconsider this function, maybe have an API call for /pools?
 export const isPoolKeyValid = (key: string) => {
   const [coinA, coinB] = key.split('-') as [CoinName, CoinName];
@@ -41,8 +44,14 @@ export const isPoolKeyValid = (key: string) => {
 
 export const createPoolIdFromPoolKey = (key: string) => {
   const [coinA, coinB, poolStability] = key.split('-') as [CoinName, CoinName, typeof StablePoolKey | typeof VolatilePoolKey];
-  const firstCoinAssetId = coinsConfig.get(coinA)?.assetId!;
-  const secondCoinAssetId = coinsConfig.get(coinB)?.assetId!;
+  const firstCoinAssetId = coinsConfig.get(coinA)?.assetId;
+  const secondCoinAssetId = coinsConfig.get(coinB)?.assetId;
+  const poolStabilityValid = poolStability === StablePoolKey || poolStability === VolatilePoolKey;
+
+  if (!firstCoinAssetId || !secondCoinAssetId || !poolStabilityValid) {
+    return null;
+  }
+
   return buildPoolId(firstCoinAssetId, secondCoinAssetId, poolStability === StablePoolKey);
 };
